Incorrect datanucleus properties used for billing boneCP
Since the original commit of the billing plot code, the datanucleus.properties file has contained a faulty configuration for its database connection library.

The BoneCP library requires properties that are named differently from those used with DBCP, which remained intact in the file.  Hence BoneCP ended up always warning that max connections were not set and that it defaulted this property to 20.

This patch replaces the faulty lines with the correct settings.

Testing:

Redeployed and restarted; no further warnings appeared in the log file.
I have tested the truncation by hacking the .jdo file to set the path datatype to 32 chars, restarting and doing a dccp to a path which exceeds that length.  A query on the database confirms the path field entry was indeed truncated and an exception was not generated.


Release Notes:  The max number of connections for the billing database is properly set to 30.  For entries into the billing tables (for VARCHAR path and errormessage), strings exceeding 8000 chars will simply be truncated without throwing an error.

Target: trunk
Patch: http://rb.dcache.org/r/4604/
Acked-by: Dmitry
Request: 2.2
Request: 2.3