ConsistentStore: fix broken unit-test
Patch http://rb.dcache.org/r/3482/, committed to trunk as r.15789 and 
back-ported to 1.9.12 as r.15790, introduced a fix to the pool's 
behaviour when subjected to a broken filesystem.  If a filesystem has a 
file yet an attempt to delete the file fails, the ConsistentStore now 
throws an RuntimeException to indicate the problem.

The patch purported to include a unit-test for this behaviour; however, 
instead of verifying what happens when the filesystem is broken, the 
test actually checked the behaviour if the file simply doesn't exist: 
an attempt to delete a non-existing file results in a RuntimeException 
being thrown.

As it happens, the behaviour the unit-test is checking is explicitly 
wrong.  This bug was fixed in a subsequent patch 
http://rb.dcache.org/r/3672/, which is committed to trunk as r.16198 
and backported to 1.9.12 with r.16199.

However, the patch fixing the bug did not also fix the broken 
unit-test.  The result is that this unit-test has been failing since 
the behaviour was fixed.

On trunk, the unit test was fixed as part of an overall rewrite of the 
test-case; see patch http://rb.dcache.org/r/4184/, committed to trunk 
as r.17000.  In that patch, the test was modified to show that no 
RuntimeException is thrown when a non-existing file is deleted.

This patch modifies the broken unit-test's behaviour to follow that of 
the refactored test.  It also includes the missing delete method on 
HFile (from r.17000) since this makes the unit-test behaviour more 
deterministic and removes the (very slight) risk of the unit-tests 
deleting files on the running computer.

Target: 1.9.12
Patch: http://rb.dcache.org/r/4607/
Acked-by: Gerd Behrmann