cells: Fix several minor issues in RoutingManager
Fixed a couple of minor performance issues: One array copy
without providing an array size, and one copy to a HashSet
without providing a target size (for domains with many
well known cells this would cause repeated reallocation
of the map as it is resized internally).

As one of the fixes introduced the use of Guava's factory
methods, I decided to use those for all collections.

I cleaned up the logging to avoid the overhead of string
concatenation if logging is not enabled.

Finally, a few minor issues got fixed: There was no need for
the _args field; the help string should be final static; and
updateUpstream logs a needless error when no default route
is installed so we now skip the upstream update when there
is no defaul route.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Acked-by: Karsten Schwank <karsten.schwank@desy.de>
Patch: http://rb.dcache.org/r/4656/