script: Make dcache dump heap user id aware
jmap can only dump the heap when running as the same user
as the JVM. Not even root can use jmap if the JVM is not
running as root.

This patch adds a check to determine if dump heap is run
as the same user as the domain. If not and if the script
is run as root, then su is used to change user account
before calling jmap.

The patch needs the whoami utility. It does not seem to be
readily available on Solaris. Therefore our script skips the
test if the whoami utility is not found.

Release notes: Fixes a permission problem when running 
dcache dump heap on domains not owned by root.

Target: trunk
Request 2.3
Request: 2.2
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4684/