pool: Fix regression and other issues in repository healer
ConsistentStore was refactored in r5062. That refactoring introduced a
bug in that BROKEN entries would not be reregistered as in FROM_CLIENT.
This means that incompletely registered uploads that fail to register
on first healing attempt will not be registered in subsequent runs of
the healer.

The bug is a consequence of the entry rebuild logic having been moved
to a separate method and thus the update of the state field to
FROM_CLIENT is lost.

This patch fixes the regression.

The patch also fixes a number of other issues:

- Healer now always recovers StorageInfo from the name space. The only
  case, in which it did not do so before, was when the entry was CACHED or
  PRECIOUS, but the file size was incorrect. I find it makes sense to
  pull the storage info from the name space in this case too.

- Some of the refactoring done in r5062 is reverted or adjusted. I do not
  believe the changes in r5062 always added to the readability. In
  particular the changes caused the abstraction level of methods to be
  heterogeneous; eg get() would call handleDelete to delete a file in one
  case, yet would delete it itself in another case. Similarly, the method
  name of isDeletable() hinted more at the recovery applied
  subsequently if the method returned true and less to why this recovery
  was applied (hasAnotherCopy() would have matched better, except it
  would not cover the DESTROYED state). In the end I believe it makes the
  code easier to read to inline this one-line method.

- I adjusted the delete logic to avoid repeated calls to delete. The
  condition !delete() && exists() is enough to catch the case in which
  a file could not be deleted and the cause is not that it didn't exist
  in the first place.

I added a unit test for the regression.

Target: trunk
Request: 2.3
Request: 2.2
Acked-by: Paul Millar <paul.millar@desy.e>
Patch: http://rb.dcache.org/r/4674/