gplazma: Let make 0 a valid value for UID and GID in kpwd file
When building dCache as root the populate script of the system-test
package fails. This is because a UID of 0 is rejected by the kpwd
utility.

The script also fails if run by a user without a grid certificate.

Both issues are fixed by this patch.

Release notes: 'dcache kpwd' commands now accept UID and GID 0 when 
creating or modifying accounts.

Target: trunk
Request: 2.3
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4689/