cells: Add request folding to RoutingManager
RoutingManager updates are relatively expensive to process: Assuming
optimal conditions for the use of hash tables, the complexity is
still linear in the number of well known cells in the downstream
domain.

For downstream domains with a large number of wellknown cells (as
would happen for busy door domains) and lots of updates (as would
happen for door domains with lots of shortlived clients), there is
a risk of accumulating a backlog of update messages.

This patch utilizes a simple property of route update messages: As
each contains a full snapshot of the well known cells of the
downstream domain, only the last update message needs to be processed.
Intermediate messages may be discarded.

The patch also ensures that RoutingManager stays response to other
requests messages.

The patch addresses part of the problem recently observed at FNAL
(although I believe the queueing of update messages is really a
symptom of another problem).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4658/