Fix array copy in IoDoorInfo
The IoDoorInfo makes a defensive copy of the array it gets as
an argument to setIoDoorEntries. Yet it proceeds by using the
original array. IntelliJ identified this as a likely bug and
I agree.

This patch fixes the bug. The patch also makes IoDoorEntry's
fields final, thus showing that we do not need to make a deep
copy of the array.

Release notes: Fixed a potential race condition in doors.

Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4731/