Fix potential deadlocks and races in CellPath
CellPath is clearly written to be thread-safe. Yet several methods
were not synchronized, thus leading to race conditions should the
class actually be used in a multi-threaded context.

The hashCode() method did not produce particularly good results. I
replaced it with the List's hash code.

I removed the support for mark() and reset() as it was not used.
Removing that feature may make it easier to turn CellPath into an
immutable class in the future.

The equals() method suffered from a deadlock (there was actually a
comment stating so). Short of making _list immutable, I could not
find a better solution that to copy the content on equals.

I suspect insert(String) is buggy in that it appears to reverse
the path during insertion. I did not fix that problem.

Require-book: no
Require-notes: no
Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4734/