chimera: fix NPE if file has no AccessLatency
JdbcFs allows a file not to have an AccessLatency or RetentionPolicy.  
Under these circumstances, a call to getAccessLatency will return 
null.

The ChimeraNamespaceProvider does not consider the possibility of a 
file having no AccessLatency or RetentionPolicy, so assumes a call to 
getAccessLatency or getRetentionPolicy will always returns a valid 
object.  It's attempt to access a method from a returned object 
results in an NPE if the file really has no AL or RP.

This has been observed during the S2 tests of 2.3.0, resulting in the 
following backtrace:

12 Jul 2012 10:24:31 (PnfsManager) [v2:srmPrepareToGet:76378220:-2136801261:-2136801260 SRM-vm-dcache-se PnfsGetFileAttributes] Error while retriving file attributes: null
java.lang.NullPointerException: null
        at org.dcache.chimera.namespace.ChimeraNameSpaceProvider.getFileAttributes(ChimeraNameSpaceProvider.java:692) ~[dcache-chimera-2.3.0-SNAPSHOT.jar:2.3.0-SNAPSHOT]
        at org.dcache.chimera.namespace.ChimeraNameSpaceProvider.getFileAttributes(ChimeraNameSpaceProvider.java:799) ~[dcache-chimera-2.3.0-SNAPSHOT.jar:2.3.0-SNAPSHOT]
        at diskCacheV111.namespace.PnfsManagerV3.getFileAttributes(PnfsManagerV3.java:1935) [dcache-core-2.3.0-SNAPSHOT.jar:2.3.0-SNAPSHOT]
        at diskCacheV111.namespace.PnfsManagerV3.processPnfsMessage(PnfsManagerV3.java:1727) [dcache-core-2.3.0-SNAPSHOT.jar:2.3.0-SNAPSHOT]
        at diskCacheV111.namespace.PnfsManagerV3$ProcessThread.run(PnfsManagerV3.java:1535) [dcache-core-2.3.0-SNAPSHOT.jar:2.3.0-SNAPSHOT]
        at java.lang.Thread.run(Thread.java:662) [na:1.6.0_33]

This patch prevents the NPE by refraining from setting the 
AccessLatency or RetentionPolicy if the file has no such attribute.  
Since both AL & RP are required values in dCache, this will likely 
result in the calling cell failing when attempting to fetch AL or RP 
value from PnfsManager's reply, but this should give more information 
about what is going wrong.

Target: trunk
Request: 2.3
Request: 2.2
Request: 1.9.12
Patch: http://rb.dcache.org/r/4647/
Acked-by: Gerd Behrmann