Fix PnfsManagerTest unit tests
To my big suprise, PnfsManagerTest.testGetAlAndRpWhenMissing is failing. This despite the fact that it passed when I ran it in IntelliJ. Turns out it only fails when run in the console.

The problem is that the test class doesn't reinitialize chimera between each test. Thus the state carries over between tests. Consequently the outcome of the tests depends on the order the tests were run in.

This also explains why testGetStorageInfoForFlushedFiles sometimes failed, or rather why it sometimes passed. Because the test should fail as it didn't set the AL and RP and thus storage info lookup fails.

Require-book: no
Require-notes: no
Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4791/