unit-tests: fix IPMatcherTest
Some of the IPMatcherTest unit-tests make use of the networking façade 
NetworkReality to provide a reliable test.  This is a static field in 
IPMatcher, so adjusting this will affect subsequent tests.  This is 
fine if all tests that do DNS activity through NetworkReality mock 
this class; however the testHostWildcardByIp does a DNS lookup without 
mocking NetworkReality.

This used to work-by-chance: since it was the first test in the class 
it, it had a good chance of being run before any of the tests that 
mock NetworkReality; however, something has changed (perhaps the 
introduction of Java 7) that resulted in the order of the tests 
changing.  The result is that the testHostWildcardByIp test is now run 
after a test that mocks NetworkReality.

This patch fixes this patch by explicitly mocking NetworkReality.  
This fix also removes the test's dependency on DNS.

Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Patch: http://rb.dcache.org/r/4796/
Acked-by: Gerd Behrmann