unit-test: ignore broken unit test
The The org.dcache.services.billing.plot.TimeFrameTest test contains a 
test called testTimeFrameDAY_MONTH2.

This randomly fails, as seen in the 2.2 branch.  Build #48 shows the 
problem.

http://svn.dcache.org/build/view/Server%20Builds/job/dCache-2.2/48/testReport/junit/org.dcache.services.billing.plot/TimeFrameTest/testTimeFrameDAY_MONTH2/

An additional problem (OOM) prevented the subsequent manually triggered 
build (#49) from building cleanly, but once that is fixed, build #50 
shows that the problem with this unit-test is transitory.

This unit test needs to be fixed or removed.

As a short-term solution, this patch marks that the test should be 
ignored; a RT ticket will track the longer-term fixing of this 
unit-test

Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Patch: http://rb.dcache.org/r/4799/
Acked-by: Gerd Behrmann