unit-tests: ignore broken unit-test
There is a unit-test for the info service that tries to check that 
metric expiry times are handled correctly.

Unfortunately, this test is dependent on the test machine's performance 
and, when under load, will be unreliable and subject to false failures.

This patch @ignores the test, pending the problem being resolved.  An 
RT ticket is created to track this issue.

Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Request: 1.9.12
Acked-by: Tigran Mkrtchyan (verbally)
Patch: http://rb.dcache.org/r/4810