[PATCH 1/2] logback: Fix merge disaster
When merging the system-test package, I intended to add a logback.xml
config just for the system-test package. However because of lack
of support for git patches in SVN, I ended up applying the changes
to the logback.xml shipped with dCache.

Thus trace and events logging ended up being enabled by default.

This patch reverts the change to the default logback.xml file and
also installs a copy with a config specific for the system-test
package.

Note that the trace and events files after this patch are once
again stored in /tmp. Moving them to the log directory causes permission
problems (as described in a recent ticket).

Alternate patch to http://rb.dcache.org/r/4809/

Target: trunk
Request: 2.3
Require-notes: yes
Require-book: no