revert patch that ignores unit-test
A previous patch, committed as 17612, commented out a unit-test as this 
patch was discovered to be broken in the 2.2 branch.  This turned out
to be a mistake, as the test is functional in the current trunk branch.

This patch reverts that change, allowing the test to run again.

Reverts: 17612@trunk
Target: Trunk