pool: Fix CDC context - again
The recent patch to avoid erroneous context information for HTTP and
xrootd transfers is ineffective. I have no idea why it worked when
I originally tested it and why it doesn't work now.

However, looking through all this code, I realized that CDCThreadFactory
is complete nonsense. This wrapper is used as a thread factory for
thread pools. The wrapper maintains the context that was valid at
the point a thread is created, however the whole point of a thread pool
is to share the threads for many requests. Thus the CDC needs to be
set per request, not per thread. Indeed, all places in which CDCThreadFactory
is used inject the CDC per request.

This patch fixes CDCThreadFactory such that it only binds the threads to
a particular cell. No other diagnostic context is initialized.

The patch changes the Netty server base class in pools so that it uses
CDCThreadFactory. This at least gets rid of the erroneous NDC. The code
is still wrong as the Netty server instance is actually shared between
all pools in a domain, whereas the logging ends up being assigned to only
one of them. Fixing this however requires changing how we instantiate
the netty server and is beyond the scope of this patch.

From the admin's point of view, this patch fixes logging of erroneous 
context information in pools for HTTP and xrootd transfers. The problem 
was attempted fixed in an earlier release, however the prior fix didn't 
work as intended.

Require-book: no
Require-notes: yes
Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4749/