xrood: Fix root path interpretation
The xrootd door allows the root path (the name space export point) to be
defined. The documentation states that the property applies to the door
in all cases, however in reality it was only used for anonymous
sessions. For authentication sessions the users private root path was
used.

There is the choice of either changing the documentation or the
implementation. I choose to change the implementation for two reasons:

- The SRM generates xrootd URIs based on the assumption that the root
  path is fixed for xrootd doors, ie that is doesn't change for each
  user.

- Personal preference that I want URIs to refer to the same resource no
  matter who uses them.

Even though the patch changes the interpretation of properties, I request
a merge to stable branches, as the bug is a potential security risk.

Require-book: no
Require-notes: yes
Target: trunk
Request: 2.3
Request: 2.2
Request: 2.1
Request: 1.9.12
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4756/