webdav: Get rid of buffering when proxying files and on directory listing
An old workaround to get rid of the buffering was invalidated when the
interpretation of the buffering property was fixed in Milton 1.6.5.

This patch updates the property so that we once again do not create
temporary buffer files in /tmp on redirect and on directory listing.


Target: trunk
Request: 2.3
Request: 2.2
Require-book: no
Require-notes: yes
Patch: http://rb.dcache.org/r/4855/
Acked-by: Tanaj Baranova <tanja.baranova@desy.de>