Expose flag for sending cache hit messages as a poolmanager property.
In the original plotting setup at Fermilab which we attempted to integrate into dCache as the billing plots component, two histogram types were provided for but never implemented:  cache hits/misses, and pool cost.   The message-passing calls for the former are already in place in the RequestContainerV5 class, but the property to turn them on was never exposed.

This patch creates a poolmanager.cache-hit-messages.enabled property and passes it down through the batch file to the spring configuration.  The default is left at false, so the property must be explicitly set to true if the administrator wishes to see this data plotted on the billing plots page, e.g.:

[dCacheDomain]
  poolmanager.cache-hit-messages.enabled=true
[dCacheDomain/poolmanager]
[dCacheDomain/broadcast]
[dCacheDomain/loginbroker]
[dCacheDomain/topo]

Target: trunk
Patch:  http://rb.dcache.org/r/4859
Request: 2.2
Request: 2.3

Require-notes: yes
Require-book: no

Acked-by: Tigran