gplazma: log if plugin doesn't instantiate or gplazma.conf is malformed
When gPlazma is started, it attempts to read the gplazma.conf file and 
instantiate configured plugins.  When a login request is received, the 
gplazma.conf file will be checked for changes.  If changes any are 
found then the configuration is re-read and the plugins reloaded.  If 
there is a syntax error with gplazma.conf or a plugin fails to 
initialise then any subsequent login attempt will trigger an attempt 
to read gplazma.conf or an attempt to load the plugins.

gPlazma will attempt to load the plugins when starting up.  If this 
fails then it is logged.  However, if the attempt to re-read 
gplazma.conf or the loading of plugins is triggered by a login request 
then only the output from the result printed is logged.  Since the 
problem appears before gPlazma starts the first phase (the auth 
phase), the problem is not recorded in the LoginResult and the logged 
information provides no hint to what went wrong.

Because the error message appears when gPlazma is started, it may not 
be clear that the logged problem will prevent any subsequent logins.  
Also, if a change to gplazma.conf triggers the problem then no 
information about the problem will be logged.

This patch fixes the problem by changing the behaviour when a login 
fails.  The code now checks if the login process started; if so then 
the normal result tree is printed.  If the login process failed to 
start then the message from the AuthenticationException is printed 
instead of the LoginResult.

FOR RELEASE NOTES:

gPlazma now reports when a login fails due to a plugin failing to 
instantiate or misconfigured gplazma.conf file.

Target: trunk
Request: 2.3
Request: 2.2
Require-notes: yes
Require-book: no