Introduce configuration cache
All our scripts start out by invoking the boot loader to compile the
configuration files. This adds considerable overhead, as starting the
JVM and compiling the configuration files takes a bit over 1 second.

I experimented with making a dedicated configuration file compiler
that avoided loading any unrelated classes, and which also optimized
various functions of the boot loader. The best I managed to do was to
bring the compilation time down from 1.3 secons to 0.8. A decent
improvement, yet still a lot of latency to add to commands that are
executed interactively.

This patch takes a different approach: Cache the oracle function. That
way the boot loader doesn't need to be invoked. I added logic to allow
the validity of the cache to be checked (compare cache age with the
age of the source files and source directories).

From the user's point of view, invocation of most of our shell scripts 
has become faster. On the downside there is a risk that cache 
invalidation fails and that /var/lib/dcache/config/cache would need to 
be deleted manually OR that /etc/dcache/dcache.conf needs to be touched.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4865/