poolmanager: Honor onerror setting and inform billing on suspend
The onerror setting was not honored in all cases. Pool manager would
choose to suspend a request on error even though it was configured to
fail. This has now been fixed.

There are still cases in which pool manager suspends requests even
though onerror is set to fail. These are when the HSM explicitly
requests a suspend, or when pool manager is configured to always
suspend certain types of requests.

The patch also adds a number of missing billing notifications. Pool
manager is supposed to inform billing when it suspends files, however
this did not happen in all cases.


Require-notes: yes
Require-book: no
Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/4775/