Change empty-string default for PnfsId to NULL
See http://rb.dcache.org/r/4833.

Setting the default value for the PnfsId in billing data was a generalization from the earlier code (1.9.12).   This really does not make sense, however.  If the data recorded from a mover or door lacks this id, it should be recorded as NULL.

This patch should be applied to trunk but not back-ported, since it is an implicit policy change.

Target: trunk
Patch: http://rb.dcache.org/r/4834
Acked-by: Dmitry
Require-notes: yes
Require-book: no