Remove CostInfo classes, methods, and tables.
In the original plotting setup at Fermilab which we attempted to integrate into dCache as the billing plots component, two histogram types were provided for but never implemented:  cache hits/misses, and pool cost.  The latter was most likely intended to convey load-balancing information and indicate hot pools to the administrator.

However, after duly considering a number of alternatives for how the latter data could be conveyed, Dmitry and I have come to the conclusion that the graphs would be redundant (if not unreadable).  Some if this information is already available from other tables that are produced in connection with read/write activity.

Hence we have judged it best to eliminate this slice of code from dCache.

This patch mainly deletes classes (PoolCostInfoMessage, PoolCostData, CostDaily) and associated methods for producing these plots in other classes (there is also a bit of incidental cleanup in a few classes as well).  The webadmin markup has also been changed so that the cost plot slots are no longer there.

Only one addition is made in this patch, that being to the changelog.

Tests: redeployed after changes, and checked logs, webadmin pages as well as database.

Target: trunk
Patch: http://rb.dcache.org/r/4856
Acked-by: Dmitry
Require-notes: yes
Require-book: no