XACML plugin does not accept x509 proxy without VO extensions.
In adapting the XACML plugin from gPlazma1 to gPlazma2, I modified the code under the assumption that the plugin should fail if no VO extensions are found in the x509 certificate chain.  This assumption, it turns out, is incorrect.  At Fermi, for instance, the GUMS server is set up to accept subject DNs without roles (FQANs) in the case that the user belongs to certain ACL groups (e.g., "fgtest").  In addition, the GsiTunnel used in connection with dcap never adds the public certificates to the Subject (only the DN Principal), in which case gsidccp would never work with the XACML plugin.

This patch fixes this bug; specifically, it allows the XACML client to request a mapping when:

1.  an x509 chain contains no extensions;
2.  the Subject contains only a DN and no public certificate chain.

In addition, one utility method originally in the x509Utils class of gPlazma1, and which explicitly parses out into simple names any fqans with "/Role=NULL" or "/Capability=NULL", has been restored to the CertificateUtils class.

Note that a subsequent patch intends to move these utility methods out of the gPlazma2 base module and into a separate dcache-gsi module, since they are specific to gsi authentication.

Tested against an installation running gsidcap and srm doors. The test user belongs to a test group ("fgtest"), but also has a VOMS fqan which maps to a community account (fnalgrid). The XACML plugin was seen to do the following:

x509 without extensions:           srmcp:    fgtest
x509 without extensions:           gsidccp:  fgtest
x509 with VOMS extensions:      srmcp:    fnalgrid
x509 with VOMS extensions:      gsidccp:  fgtest

An aside: it may be useful to review why the GsiTunnel has a special path through authentication and does not provide the Login module with a Subject containing the certificate chain as well. 

Target: trunk
Patch: http://rb.dcache.org/r/4902
Request: 2.2
Request: 2.3
Require-notes: yes
Require-book: no
Acked-by: Dmitry