X509/VOMS Utility refactoring
This patch refactors the GSI-specific code  in the org.dcache.gplazma.utils package in the gPlazma2 module as follows:

1.  A new module, dcache-gsi, has been created.

2.  The CertificateUtils methods have been broken up into X509 and GSS utility classes and shifted out of the gPlazma2 base module, since they are protocol-specific, and into the gsi utility module.

3.  The dependencies on the new methods and module have been substituted in the relevant classes and pom.xml files.

4.  NameRolePair has been shifted to the gplazma-grid module where it belongs.

5.  Wherever the PKIVerifier is used, this has been turned into an object-instance rather than a static instance, since in theory one might have several instances of VOMS-based plugins within a single domain with potentially different VOMS directories.  The mapping of one verifier per instance also ensures that the CA certs need not be reloaded more than necessary (but note that there exists a thread in the verifier which itself refreshes the certs every minute or so).

It should be noted that the location of the VOMS directory is configurable for the VOMS and XACML plugins, but it remains fixed (as it was originally) at the standard location for the GsiTunnel and the SRMAuthorizer.  This may need to be improved eventually.

While this patch appears to introduce numerous changes, most are of a minor nature resulting from #3 above.

Target: trunk
Patch: http://rb.dcache.org/r/4903
Acked-by: Dmitry
Require-notes: yes
Require-book: no