Logging-based alarms, part 2 (storage/DAO)
This is part 2 (out of 4) of the third version of the proposed alarms component.  

In order to persist alarm messages, an AlarmEntry DAO is used.  All alarms have the same structure and thus are uniformly mapped to this object.  An API is defined for writing these objects to an underlying store, implemented using a DataNucleus layer.  The plugin currently chosen to handle the actual on-disk storage is the DataNucleus XML store.  Given that alarms should under normal circumstances not be high-traffic/volume, this non-RDBMS plugin should be sufficient and lightweight enough for the purpose. It should be noted, however, that the third patch in this series also provides automatic clean-up as part of the front-end (http) service configuration which displays the alarms.

A logback appender which converts Alarm events into the AlarmEntry storage type and calls put on the storage API is also provided.  The way this may be used is illustrated by the logback-server.xml meant to configure the remote logging service which will receive the events;  also included there is the skeleton for configuring an SMTP appender for forwarding the alarms to email recipients:

    <!--
        this appender will intercept all logging messages sent with the
        ALARM marker and store them in the alarms (XML) database
    -->
    <appender name="ALARM" class="org.dcache.alarms.logback.AlarmEntryAppender">
        <!-- this filter ensures that only events sent marked as ALARM
             are received by this appender -->
        <filter class="org.dcache.alarms.logback.AlarmMarkerFilter"/>
    </appender>

    <!--
        if you wish alarms to be sent as email, you will need to provide host,
        to and from information below (note the possibility of multiple 'to's),
        and then include the appender in the root logger.
    -->
    <appender name="ALARM_MAIL" class="ch.qos.logback.classic.net.SMTPAppender">
        <!-- this filter ensures that only events sent marked as ALARM
             are received by this appender -->
        <filter class="org.dcache.alarms.logback.AlarmMarkerFilter"/>
        <smtpHost></smtpHost>
        <to></to>
        <to></to>
        <from></from>
        <subject>dCache Alarm</subject>
        <layout class="ch.qos.logback.classic.PatternLayout">
            <pattern>%d{dd MMM yyyy HH:mm:ss} \(%X{cells.cell}\) [%X{org.dcache.ndc}] %m%n</pattern>
        </layout>
        <cyclicBufferTracker class="ch.qos.logback.core.spi.CyclicBufferTrackerImpl">
            <!-- send just one log entry per email -->
            <bufferSize>1</bufferSize>
        </cyclicBufferTracker>
    </appender>

NOTE that two additional elements, <xmlPath> and <propertiesPath>, are available on the AlarmEntryAppender; these are for overriding the default location of the xml store file (/var/lib/dcache/alarms/alarms.xml) and for providing a custom implementation of (configuration of) the underlying database (cf. http://rb.dcache.org/r/4890, "webadmin.alarm.dao.properties).  Should either of these be modified from the default values, care must be taken to configure the webadmin alarms service to match the changes.

Also included is a bash script for launching the server itself, defined against a classpath which accesses the dCache installation (/usr/shared/dcache/classes).  It is understood that the server will run on top of a file system shared by the httpd service, since these will both access the storage file generated at /var/lib/dcache/alarms/alarms.xml.

This patch supersedes http://rb.dcache.org/r/4663; much of the code is unchanged.

Target: trunk
Patch: http://rb.dcache.org/r/4888
Acked-by: Dmitry
Require-notes: yes
Require-book: yes